Send status 422 when missing parameter

+++ backend/src/controller/department/management/removeCentre.js
@@ -6,6 +6,11 @@ import pool from "../../../configs/connectDB";
 
 let removeCentre = async (req, res) => {
     let id = req.body.id
+
+    if (id === undefined) {
+        return res.status(422).send({message: 'Missing parameter!'})
+    }
+
     let query = 'delete from centre where id = ?; delete from account where id = ?'
     await pool.query(query, [id])
     return res.send({message: 'Centre removed'})
